Don't always scroll to bottom
If window isn't scrolled to bottom, don't auto scroll on new element.
This help when user wants to keep focused on some line in the middle of the
log window (like when reading messages in Whatsapp).
Keep auto scrolling to bottom if already on bottom.

Also, add minimal height to line, this help recognize empty lines (like empty
string)